Шаблонизация проекта (v.2.0)
